Use XCB_CURRENT_TIME for randr function calls. Fix focus on ws change.
Fix focus handling when changing workspace. Before when changing
workspace the focus would be on the window where the mouse is
regardless if that was the focus window or not. The fix is rather
crude and just maintains a global counter of events to ignore but
so far it seems to work well.