Fix focus change on workspace switch, large cleanup.
When switching workspaces FocusIn events are generated, causing the
focus to be grabbed by the client under the mouse pointer. This approach
simply flushes the server then ignores all queued events until the queue
is empty, we may lose some minor events but this all happens so fast it
likely won't matter.