Fix: error using root window depth for client borders.
Store the client window depth at initialization for use later.
The pixmap and the window it is being mapped must have the same
depth or the operation will fail. This may still pose a problem
later if the client changes it's depth and we aren't aware of it
but we'll deal with that if it becomes a problem.