Fix: crash on commands being issued during move/resize.
Because we're using select() to multiplex IO on the command fd and X
connection fd we cannot have separate event loops *(like we had in
mousemvr())*. I'm still not certain this solves potential crashes caused
by this but so far so good.

This moves all the code for motion notify, button press, and button
release events into eventhandle(). Also inlined grabpointer() as it's
only called for button press events when starting a move or resize.