Fix up the previous change to layout definition
This was a quick fix but not very nice, instead we keep a separate
internal array of the initial layout names and use them to match,
because it's a 1-1 mapping we can be sure the indexing will be correct.
It also uses less memory and should be the same performance if not
better.